% Generated manually for the initial package scaffold.
\name{hector_engine_parameters}
\alias{hector_engine_parameters}
\title{Return HECTOR engine parameters}
\description{Returns the main field and probe geometry parameters of a configuration engine.}
\usage{
hector_engine_parameters(engine = .hector_new_engine())
}
\arguments{\item{engine}{A private engine environment.}}
\value{A named list of field and probe parameters.}
